Move requestDebuggingEnabled to IDebugger.isAvailable
