+++ tests/src/session.spec.ts
@@ -287,47 +287,3 @@ describe('protocol', () => {
     });
   });
 });
-
-describe('DebugSession statics', () => {
-  let xpythonClient: IClientSession;
-  let ipykernelClient: IClientSession;
-
-  beforeAll(async () => {
-    xpythonClient = await createClientSession({
-      kernelPreference: {
-        name: 'xpython'
-      }
-    });
-    ipykernelClient = await createClientSession({
-      kernelPreference: {
-        name: 'python3'
-      }
-    });
-    await Promise.all([
-      (xpythonClient as ClientSession).initialize(),
-      (ipykernelClient as ClientSession).initialize()
-    ]);
-    await Promise.all([
-      xpythonClient.kernel.ready,
-      ipykernelClient.kernel.ready
-    ]);
-  });
-
-  afterAll(async () => {
-    await Promise.all([xpythonClient.shutdown(), ipykernelClient.shutdown()]);
-  });
-
-  describe('#requestDebuggingEnabled', () => {
-    it('should return true for kernels that have support for debugging', async () => {
-      const enabled = await DebugSession.requestDebuggingEnabled(xpythonClient);
-      expect(enabled).to.be.true;
-    });
-
-    it('should return false for kernels that do not have support for debugging', async () => {
-      const enabled = await DebugSession.requestDebuggingEnabled(
-        ipykernelClient
-      );
-      expect(enabled).to.be.false;
-    });
-  });
-});
